[CLEANUP] Last bit of ui elements (#4472)
* moving relation status fill bar

* moving relation status scale bar

* moving relation display

* moving dropdown container

* Update ui_elements.py

* moving checkbox

* moving cat list display

* moving modified image

* moving dropdown

* move collapsible container

* Delete ui_elements.py

* fix merge
