--- scripts/game_structure/ui_elements.py
@@ -1,1381 +1,0 @@
-import math
-from functools import lru_cache
-from math import ceil
-from typing import (
-    Optional,
-    List,
-    Union,
-    Dict,
-)
-
-import i18n
-import pygame
-import pygame_gui
-from pygame_gui.core import UIContainer, IContainerLikeInterface, UIElement, ObjectID
-from pygame_gui.core.gui_type_hints import RectLike
-from pygame_gui.core.interfaces import IUIManagerInterface
-from pygame_gui.elements import UIAutoResizingContainer
-
-from scripts.cat_relations.enums import RelType, RelTier
-from scripts.clan_package.settings import get_clan_setting
-from scripts.game_structure import image_cache
-from scripts.game_structure import game
-from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.game.settings import game_setting_get
-from scripts.ui.elements.image_button import UIImageButton
-from scripts.ui.elements.sprite_button import UISpriteButton
-from scripts.ui.elements.surface_image_button import UISurfaceImageButton
-from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
-from scripts.ui.generate_box import BoxStyles, get_box
-from scripts.ui.generate_button import get_button_dict, ButtonStyles
-from scripts.ui.icon import Icon
-from scripts.events_module.text_adjust import shorten_text_to_fit
-from scripts.ui.scale import ui_scale, ui_scale_dimensions, ui_scale_value
-
-
-class UIRelationStatusFillBar(pygame_gui.elements.UIStatusBar):
-    """Wraps together a status bar"""
-
-    def __init__(
-        self,
-        relative_rect,
-        percent_full=0,
-        manager=None,
-        anchors=None,
-        tool_tip_text: str = None,
-        container=None,
-    ):
-        rect = (
-            (relative_rect.x + ui_scale_value(2), relative_rect.y + ui_scale_value(2)),
-            (
-                relative_rect.width - ui_scale_value(4),
-                relative_rect.height - ui_scale_value(4),
-            ),
-        )
-        super().__init__(
-            rect,
-            object_id="#relation_bar",
-            manager=manager,
-            anchors=anchors,
-            container=container,
-        )
-        self.percent_full = percent_full / 100
-
-        # Now to make the overlay
-        image = pygame.transform.scale(
-            image_cache.load_image(
-                "resources/images/relations_border_bars.png"
-            ).convert_alpha(),
-            (relative_rect[2], relative_rect[3]),
-        )
-
-        self.overlay = pygame_gui.elements.UIImage(
-            relative_rect,
-            image,
-            manager=manager,
-            anchors=anchors,
-            object_id="#relation_bar",
-            container=container,
-        )
-        self.overlay.set_tooltip(tool_tip_text)
-        self.overlay.tool_tip_delay = 0
-        self.join_focus_sets(self.overlay)
-
-    def kill(self):
-        self.overlay.kill()
-        super().kill()
-
-
-class UIRelationStatusScaleBar(pygame_gui.elements.UIImage):
-    """Wraps together a status bar"""
-
-    def __init__(
-        self,
-        relative_rect,
-        tier: RelTier,
-        container=None,
-        manager=None,
-        anchors: dict = None,
-        starting_height: int = 1,
-        scale_position: int = 0,
-        tool_tip_text: str = None,
-    ):
-        # creating the colored bar
-        path = "resources/images/relation_bar.png"
-
-        bar = pygame.transform.scale(
-            image_cache.load_image(path),
-            (relative_rect[2], relative_rect[3]),
-        )
-
-        bar_center_offset = int(bar.width / 2)
-        num_sub_bars = math.ceil((abs(scale_position)) / 25)
-        sub_bar_width = bar.width // 8
-        bar_width = sub_bar_width * num_sub_bars
-
-        if scale_position < 0:
-            short_bar = pygame.Rect(
-                bar_center_offset - bar_width, 0, bar_width, bar.height
-            )
-        else:
-            short_bar = pygame.Rect(bar_center_offset, 0, bar_width, bar.height)
-
-        surf = pygame.Surface((short_bar.width, short_bar.height))
-
-        bar.fill((130, 117, 82))
-
-        bar_colour = (130, 117, 82)
-        if tier.is_low_pos:
-            bar_colour = (182, 174, 51)
-        elif tier.is_mid_pos:
-            bar_colour = (150, 195, 49)
-        elif tier.is_extreme_pos:
-            bar_colour = (154, 241, 32)
-        elif tier.is_low_neg:
-            bar_colour = (186, 128, 60)
-        elif tier.is_mid_neg:
-            bar_colour = (214, 90, 53)
-        elif tier.is_extreme_neg:
-            bar_colour = (233, 38, 30)
-
-        surf.fill(bar_colour)
-        bar.blit(surf, (short_bar.left, short_bar.top))
-
-        # bar element is the base of this entire element
-        super().__init__(
-            relative_rect,
-            bar,
-            container=container,
-            manager=manager,
-            starting_height=starting_height,
-            anchors=anchors,
-        )
-
-        # Now to make the overlay
-        image = pygame.transform.scale(
-            image_cache.load_image(
-                "resources/images/relations_border_bars.png"
-            ).convert_alpha(),
-            (relative_rect[2], relative_rect[3]),
-        )
-
-        # overlay element
-        self.overlay = UIModifiedImage(
-            relative_rect,
-            image,
-            manager=manager,
-            container=container,
-            anchors=anchors,
-            starting_height=starting_height,
-            object_id="#relation_bar",
-        )
-        self.join_focus_sets(self.overlay)
-        self.overlay.set_tooltip(tool_tip_text)
-        self.overlay.tool_tip_delay = 0
-
-        # pointer element
-        # -7 bc coords are top-left so we have to shift over so arrow points at middle
-        pointer_origin = (bar.width // 2 - ui_scale_value(7), 0)
-        # every "unit" is 1/200th of the width of the bar
-        pointer_offset = int(scale_position / 200 * bar.width)
-        # -15 so it doesn't go past the end of the bar
-        pointer_x = max(
-            0, min(pointer_offset + pointer_origin[0], bar.width - ui_scale_value(15))
-        )
-
-        pointer_final_position = (
-            pointer_x,
-            pointer_origin[1],
-        )
-        pointer_size = ui_scale_dimensions((14, 12))
-
-        pointer = pygame.transform.scale(
-            image_cache.load_image("resources/images/rel_pointer.png").convert_alpha(),
-            pointer_size,
-        )
-        self.pointer = pygame_gui.elements.UIImage(
-            pygame.Rect(pointer_final_position, pointer_size),
-            pointer,
-            manager=manager,
-            container=container,
-            anchors=anchors,
-            starting_height=starting_height,
-        )
-        self.join_focus_sets(self.pointer)
-
-    def kill(self):
-        self.overlay.kill()
-        self.pointer.kill()
-        super().kill()
-
-
-class UIRelationDisplay(pygame_gui.elements.UIAutoResizingContainer):
-    def __init__(
-        self,
-        position: tuple,
-        relationship,
-        romance: bool = False,
-        container=None,
-        manager=None,
-        anchors=None,
-        starting_height=1,
-    ):
-        dimensions = (0, 0)
-        self.rel_elements = {}
-        bar_size = (96, 10)
-
-        super().__init__(
-            relative_rect=ui_scale(pygame.Rect(position, dimensions)),
-            container=container,
-            manager=manager,
-            anchors=anchors,
-            starting_height=starting_height,
-        )
-
-        prev_element = None
-        for rel_type in [*RelType]:
-            if rel_type == RelType.ROMANCE:
-                continue
-            num, tier = relationship.get_rel_type_attributes(rel_type)
-            self.rel_elements[f"{rel_type}_text"] = pygame_gui.elements.UITextBox(
-                f"relationships.{tier}",
-                ui_scale(
-                    pygame.Rect(
-                        (0 - 2, 0),
-                        (100, 25),
-                    )
-                ),
-                object_id="#text_box_26_horizcenter",
-                container=self,
-                anchors={"top_target": prev_element} if prev_element else None,
-            )
-            self.rel_elements[f"{rel_type}_text"].set_tooltip(
-                i18n.t(f"relationships.{rel_type}", count=num)
-            )
-            self.rel_elements[f"{rel_type}_text"].tool_tip_delay = 0
-            self.rel_elements[f"{rel_type}_text"].disable()
-            self.rel_elements[f"{rel_type}_bar"] = UIRelationStatusScaleBar(
-                ui_scale(pygame.Rect((0, -5), bar_size)),
-                tier,
-                anchors={"top_target": self.rel_elements[f"{rel_type}_text"]},
-                scale_position=num,
-                container=self,
-            )
-            prev_element = self.rel_elements[f"{rel_type}_bar"]
-
-            # ROMANCE
-        if romance:
-            self.rel_elements[f"romance_text"] = UITextBoxTweaked(
-                f"relationships.{relationship.romance_tier if relationship.romance_tier else 'neutral'}",
-                ui_scale(
-                    pygame.Rect(
-                        (0, 1),
-                        (96, -1),
-                    )
-                ),
-                object_id="#text_box_26_horizcenter",
-                anchors={"top_target": prev_element},
-                container=self,
-                line_spacing=0.95,
-            )
-            self.rel_elements[f"romance_text"].set_tooltip(
-                i18n.t(f"relationships.romance", count=relationship.romance)
-            )
-            self.rel_elements[f"romance_text"].tool_tip_delay = 0
-            self.rel_elements[f"romance_text"].disable()
-            self.rel_elements[f"romance_text"].disable()
-            self.rel_elements[f"romance_bar"] = UIRelationStatusFillBar(
-                ui_scale(
-                    pygame.Rect(
-                        (0, -5),
-                        bar_size,
-                    )
-                ),
-                relationship.romance,
-                container=self,
-                anchors={"top_target": self.rel_elements[f"romance_text"]},
-            )
-
-        self.romance = romance
-
-
-class UIDropDownContainer(UIAutoResizingContainer):
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        container: UIContainer,
-        manager: IUIManagerInterface,
-        starting_height: int = 1,
-        object_id: str = None,
-        visible: bool = False,
-        anchors: dict = None,
-        child_trigger_close: bool = False,
-        starting_selection: list = None,
-        open_on_hover: bool = False,
-    ):
-        """
-        Holds all the elements of a dropdown and coordinates its basic responses.
-        :param relative_rect: The starting size and relative position of the container.
-        :param container: The container this container is within. Defaults to None (which is the root
-                          container for the UI)
-        :param starting_height: The starting layer height of this container above its container.
-                                Defaults to 1.
-        :param object_id: An object ID for this element.
-        :param manager: The UI manager for this element. If not provided or set to None,
-                        it will try to use the first UIManager that was created by your application.
-        :param visible: Whether the element is visible by default. Warning - container visibility
-                        may override this.
-        :param open_on_hover: Dropdown will open while being hovered and close once unhovered
-
-        """
-        self.open_on_hover = open_on_hover
-
-        super().__init__(
-            relative_rect=relative_rect,
-            container=container,
-            object_id=object_id,
-            starting_height=starting_height,
-            visible=visible,
-            manager=manager,
-            anchors=anchors,
-        )
-
-        self.parent_button = None
-        self.child_button_container = None
-        self.child_buttons = []
-        self.child_button_dicts = []
-
-        self.is_open: bool = False
-        self.child_trigger_close = child_trigger_close
-        self.selected_list = (
-            [item for item in starting_selection] if starting_selection else []
-        )
-
-    def close(self):
-        """
-        closes the dropdown
-        """
-        self.child_button_container.hide()
-
-        self.resize_bottom = False
-        self.set_dimensions(self.parent_button.get_relative_rect().size)
-
-        self.is_open = False
-
-    def open(self):
-        """
-        opens the dropdown
-        """
-        self.resize_bottom = True
-        self.should_update_dimensions = True
-
-        self.child_button_container.show()
-        self.is_open = True
-
-    def disable_child(self, item_name, button=None):
-        """
-        disables the given element and enables all other children
-        clears self.selected_list and adds given item_name to it
-        """
-        if not button:
-            button = self.child_button_dicts[item_name]
-
-        button.disable()
-        self.selected_list.clear()
-        self.selected_list.append(item_name)
-
-        for child in self.child_button_container.elements:
-            if child == button:
-                continue
-            child.enable()
-
-    def check_if_hovering(self):
-        mouse_x, mouse_y = self.ui_manager.get_mouse_position()
-        if self.hover_point(mouse_x, mouse_y):
-            return True
-        else:
-            return False
-
-    def update(self, time_delta: float):
-        # hover
-        if self.open_on_hover:
-            if self.check_if_hovering():
-                if not self.is_open and self.parent_button.hovered:
-                    self.open()
-            else:
-                self.close()
-        # press
-        else:
-            if self.parent_button.pressed:
-                if self.is_open:
-                    self.close()
-                else:
-                    self.open()
-
-        super().update(time_delta)
-
-
-class UICheckbox(UIImageButton):
-    """
-    Creates a checkbox and allows for easy check and uncheck
-    :param position: The relative position of the container.
-    :param container: The container this container is within. Defaults to None (which is the root
-                      container for the UI)
-    :param starting_height: The starting layer height of this container above its container.
-                            Defaults to 1.
-    :param manager: The UI manager for this element. If not provided or set to None,
-                    it will try to use the first UIManager that was created by your application.
-    :param visible: Whether the element is visible by default. Warning - container visibility
-                    may override this.
-    :param check: the checkbox begins in the "checked" state, default False
-    """
-
-    def __init__(
-        self,
-        position: tuple,
-        container: IContainerLikeInterface,
-        manager,
-        visible: bool = True,
-        tool_tip_text: str = None,
-        starting_height: int = 1,
-        check: bool = False,
-        anchors=None,
-    ):
-        self.checked = check
-
-        relative_rect = ui_scale(pygame.Rect(position, (34, 34)))
-
-        if check:
-            object_id = "@checked_checkbox"
-        else:
-            object_id = "@unchecked_checkbox"
-
-        super().__init__(
-            relative_rect=relative_rect,
-            text="",
-            container=container,
-            tool_tip_text=tool_tip_text,
-            starting_height=starting_height,
-            visible=visible,
-            manager=manager,
-            object_id=object_id,
-            anchors=anchors,
-        )
-
-    def check(self):
-        """
-        switches the checkbox into the "checked" state
-        """
-        self.checked = True
-        self.change_object_id("@checked_checkbox")
-
-    def uncheck(self):
-        """
-        switches the checkbox into the "unchecked" state
-        """
-        self.checked = False
-        self.change_object_id("@unchecked_checkbox")
-
-    def hover_point(self, hover_x: float, hover_y: float) -> bool:
-        """
-        Test if a given point counts as 'hovering' this UI element. Normally that is a
-        straightforward matter of seeing if a point is inside the rectangle. Occasionally it
-        will also check if we are in a wider zone around a UI element once it is already active,
-        this makes it easier to move scroll bars and the like.
-
-        :param hover_x: The x (horizontal) position of the point.
-        :param hover_y: The y (vertical) position of the point.
-
-        :return: Returns True if we are hovering this element.
-
-        """
-
-        container_clip_rect = self.ui_container.get_container().get_rect().copy()
-        if self.ui_container.get_container().get_image_clipping_rect() is not None:
-            container_clip_rect.size = (
-                self.ui_container.get_container().get_image_clipping_rect().size
-            )
-            container_clip_rect.left += (
-                self.ui_container.get_container().get_image_clipping_rect().left
-            )
-            container_clip_rect.top += (
-                self.ui_container.get_container().get_image_clipping_rect().top
-            )
-
-        # ONLY CHANGE was to remove the drawable shape collide point check. for some reason, it would cause the checkbox
-        # hover to desync when inside a scrolling container
-
-        return bool(self.rect.collidepoint(hover_x, hover_y)) and bool(
-            container_clip_rect.collidepoint(hover_x, hover_y)
-        )
-
-
-class UICatListDisplay(UIContainer):
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        container: UIContainer,
-        starting_height: int,
-        manager,
-        cat_list: list,
-        cats_displayed: int,
-        x_px_between: int,
-        columns: int,
-        current_page: int,
-        next_button: UIImageButton,
-        prev_button: UIImageButton,
-        object_id: str = None,
-        first_button: UIImageButton = None,
-        last_button: UIImageButton = None,
-        anchors: Optional[dict] = None,
-        rows: int = None,
-        show_names: bool = False,
-        tool_tip_text_list: list = None,
-        tool_tip_name: bool = False,
-        tool_tip_nutrition: bool = False,
-        custom_sprites_object_id: str = None,
-        visible: bool = True,
-        text_theme="#cat_list_text",
-        y_px_between: int = None,
-        allow_selection: bool = False,
-        hover_input_only: bool = False,
-    ):
-        """
-        Creates and displays a list of click-able cat sprites.
-        :param relative_rect: The starting size and relative position of the container.
-        :param container: The container this container is within. Defaults to None (which is the root
-                          container for the UI)
-        :param starting_height: The starting layer height of this container above its container.
-                                Defaults to 1.
-        :param object_id: An object ID for this element.
-        :param manager: The UI manager for this element. If not provided or set to None,
-                        it will try to use the first UIManager that was created by your application.
-        :param cat_list: the list of cat objects that need to display
-        :param cats_displayed: the number of cats to display on one page
-        :param x_px_between: the pixel space between each column of cats
-        :param y_px_between: the pixel space between each row of cats. Optional, defaults to x_px_between
-        :param columns: the number of cats in a row before a new row is created
-        :param next_button: the next_button ui_element
-        :param prev_button: the prev_button ui_element
-        :param current_page: the currently displayed page of the cat list
-        :param tool_tip_name: should a tooltip displaying the cat's name be added to each cat sprite, default False
-        :param tool_tip_nutrition: should a tooltip displaying the cat's nutrition status be added to each cat sprite, default False
-        :param visible: Whether the element is visible by default. Warning - container visibility
-                        may override this.
-        :param allow_selection: Whether cats should be selectable.
-        """
-
-        super().__init__(
-            relative_rect=relative_rect,
-            container=container,
-            starting_height=starting_height,
-            object_id=object_id,
-            visible=visible,
-            anchors=anchors,
-            manager=manager,
-        )
-
-        self.cat_list = cat_list
-        self.cats_displayed = cats_displayed
-        self.x_px_between = x_px_between
-        self.y_px_between = y_px_between if y_px_between is not None else x_px_between
-        self.columns = columns
-        self.rows = rows if rows is not None else ceil(cats_displayed / columns)
-        self.current_page = current_page
-        self.next_button = next_button
-        self.prev_button = prev_button
-        self.first_button = first_button
-        self.last_button = last_button
-        self.tool_tip_text = tool_tip_text_list
-        self.tool_tip_name = tool_tip_name
-        self.tool_tip_nutrition = tool_tip_nutrition
-        self.custom_sprites_object_id = custom_sprites_object_id
-        self.text_theme = text_theme
-        self.allow_selection = allow_selection
-
-        self.total_pages: int = 0
-        self.favor_indicator = {}
-        self.cat_sprites = {}
-        self.cat_names = {}
-        self.cat_chunks = []
-        self.boxes = []
-        self.selection_boxes = {}
-        self.selected = []
-
-        self.show_names = show_names
-
-        self._favor_circle = pygame.transform.scale(
-            pygame.image.load(f"resources/images/fav_marker.png").convert_alpha(),
-            ui_scale_dimensions((50, 50)),
-        )
-        if game_setting_get("dark mode"):
-            self._favor_circle.set_alpha(150)
-
-        self.generate_grid()
-
-        self._chunk()
-        self._display_cats()
-
-    def generate_grid(self):
-        """
-        A wrapper for the grid generation to speed it up significantly.
-        Must be done like this to avoid memory leak.
-        """
-        self.boxes = self._generate_grid_cached(
-            self.relative_rect.width // self.columns,
-            self.relative_rect.height // self.rows,
-            self.rows,
-            self.columns,
-            self.ui_manager,
-        )
-        for box in self.boxes:
-            box.set_container(self)
-            box.rebuild()
-
-    def cache_clear(self):
-        """
-        Clears the cached grid. This is only necessary for cat lists being displayed on popup windows. I'm not sure *why*, but the cache starts causing crashes. I recommend that we try to keep cat list displays on popup windows to a minimum to avoid lag and, when possible, hide & show the list instead of killing and recreating.
-        """
-
-        self._generate_grid_cached.cache_clear()
-
-    @staticmethod
-    @lru_cache(maxsize=5)
-    def _generate_grid_cached(cell_width, cell_height, rows, columns, manager):
-        boxes: List[Optional[UIContainer]] = [None] * (rows * columns)
-        for i, box in enumerate(boxes):
-            if i == 0:
-                anchors = {}
-            elif i % columns == 0:
-                # first item in a row excluding first
-                anchors = {"top_target": boxes[i - columns]}
-            elif i < columns:
-                # top row
-                anchors = {"left_target": boxes[i - 1]}
-            else:
-                # all other rows
-                anchors = {
-                    "left_target": boxes[i - 1],
-                    "top_target": boxes[i - columns],
-                }
-
-            boxes[i] = UIContainer(
-                pygame.Rect(
-                    0,
-                    0,
-                    cell_width,
-                    cell_height,
-                ),
-                anchors=anchors,
-                manager=manager,
-            )
-        return boxes
-
-    def clear_display(self):
-        [sprite.kill() for sprite in self.cat_sprites.values()]
-        [name.kill() for name in self.cat_names.values()]
-        [favor.kill() for favor in self.favor_indicator.values()]
-        self.next_button = None
-        self.prev_button = None
-        self.first_button = None
-        self.last_button = None
-
-    def update_display(self, current_page: int, cat_list: list):
-        """
-        updates current_page and refreshes the cat display
-        :param current_page: the currently displayed page
-        :param cat_list: the new list of cats to display, leave None if list isn't changing, default None
-        """
-
-        self.current_page = current_page
-        if cat_list != self.cat_list:
-            self.cat_list = cat_list
-            self._chunk()
-        self._display_cats()
-
-    def _chunk(self):
-        """
-        separates the cat list into smaller chunks to display on each page
-        """
-        self.cat_chunks = [
-            self.cat_list[x : x + self.cats_displayed]
-            for x in range(0, len(self.cat_list), self.cats_displayed)
-        ]
-
-    def _display_cats(self):
-        """
-        creates the cat display
-        """
-        self.current_page = max(1, min(self.current_page, len(self.cat_chunks)))
-
-        self._update_arrow_buttons()
-
-        display_cats = []
-        if self.cat_chunks:
-            self.total_pages = len(self.cat_chunks)
-            display_cats = self.cat_chunks[self.current_page - 1]
-
-        [sprite.kill() for sprite in self.cat_sprites.values()]
-        [name.kill() for name in self.cat_names.values()]
-        [favor.kill() for favor in self.favor_indicator.values()]
-
-        show_fav = get_clan_setting("show fav")
-
-        # FAVOURITE ICON
-        if show_fav:
-            fav_indexes = [
-                display_cats.index(cat) for cat in display_cats if cat.favourite
-            ]
-            [self.create_favor_indicator(i, self.boxes[i]) for i in fav_indexes]
-
-        # CAT SPRITE
-        [
-            self.create_cat_button(i, kitty, self.boxes[i])
-            for i, kitty in enumerate(display_cats)
-        ]
-
-        # CAT NAME
-        if self.show_names:
-            [
-                self.create_name(i, kitty, self.boxes[i])
-                for i, kitty in enumerate(display_cats)
-            ]
-
-    def create_cat_button(self, i, kitty, container):
-        if self.tool_tip_nutrition:
-            condition_list = []
-            if kitty.illnesses:
-                if "starving" in kitty.illnesses.keys():
-                    condition_list.append(i18n.t("conditions.illnesses.starving"))
-                elif "malnourished" in kitty.illnesses.keys():
-                    condition_list.append(i18n.t("conditions.illnesses.malnourished"))
-            nutrition_info = game.clan.freshkill_pile.nutrition_info
-            if kitty.ID in nutrition_info:
-                full_text = i18n.t(
-                    "screens.profile.nutrition_text",
-                    nutrition_text=nutrition_info[kitty.ID].nutrition_text,
-                )
-                if get_clan_setting("showxp"):
-                    full_text += f" ({str(int(nutrition_info[kitty.ID].percentage))})"
-                condition_list.append(full_text)
-            tooltip_text = (
-                "<br>".join(condition_list) if len(condition_list) > 0 else None
-            )
-        elif self.tool_tip_name:
-            tooltip_text = str(kitty.name)
-        elif self.tool_tip_text:
-            tooltip_text = self.tool_tip_text[i]
-        else:
-            tooltip_text = None
-
-        if self.allow_selection:
-            self.selection_boxes[f"sprite{i}"] = pygame_gui.elements.UIImage(
-                ui_scale(pygame.Rect((0, 15), (56, 56))),
-                get_box(BoxStyles.SELECTION_BOX, (60, 60)),
-                container=container,
-                starting_height=1,
-                manager=MANAGER,
-                visible=False,
-                anchors={"centerx": "centerx"},
-            )
-        self.cat_sprites[f"sprite{i}"] = UISpriteButton(
-            ui_scale(pygame.Rect((0, 15), (50, 50))),
-            kitty.sprite,
-            cat_object=kitty,
-            cat_id=kitty.ID,
-            mask=None,
-            container=container,
-            object_id=f"#cat_sprite"
-            if not self.custom_sprites_object_id
-            else self.custom_sprites_object_id,
-            tool_tip_text=tooltip_text,
-            starting_height=1,
-            anchors={"centerx": "centerx"},
-        )
-
-    def create_name(self, i, kitty, container):
-        self.cat_names[f"name{i}"] = pygame_gui.elements.UILabel(
-            pygame.Rect((0, 0), (container.rect[2], ui_scale_value(30))),
-            shorten_text_to_fit(str(kitty.name), 220, 30),
-            container=container,
-            object_id=self.text_theme,
-            anchors={
-                "centerx": "centerx",
-                "top_target": self.cat_sprites[f"sprite{i}"],
-            },
-        )
-
-    def create_favor_indicator(self, i, container):
-        self.favor_indicator[f"favor{i}"] = pygame_gui.elements.UIImage(
-            ui_scale(pygame.Rect((0, 15), (50, 50))),
-            self._favor_circle,
-            object_id=f"favor_circle{i}",
-            container=container,
-            starting_height=1,
-            anchors={"centerx": "centerx"},
-        )
-
-    def _update_arrow_buttons(self):
-        """
-        enables/disables appropriate arrow buttons
-        """
-        if len(self.cat_chunks) <= 1:
-            self.prev_button.disable()
-            self.next_button.disable()
-            if self.first_button:
-                self.first_button.disable()
-                self.last_button.disable()
-        elif self.current_page >= len(self.cat_chunks):
-            self.prev_button.enable()
-            self.next_button.disable()
-            if self.first_button:
-                self.first_button.enable()
-                self.last_button.disable()
-        elif self.current_page == 1 and len(self.cat_chunks) > 1:
-            self.prev_button.disable()
-            self.next_button.enable()
-            if self.first_button:
-                self.first_button.disable()
-                self.last_button.enable()
-        else:
-            self.prev_button.enable()
-            self.next_button.enable()
-            if self.first_button:
-                self.first_button.enable()
-                self.last_button.enable()
-
-    def process_event(self, event: pygame.event.Event) -> bool:
-        if self.allow_selection and event.type in (
-            pygame_gui.UI_BUTTON_ON_HOVERED,
-            pygame_gui.UI_BUTTON_ON_UNHOVERED,
-            pygame_gui.UI_BUTTON_START_PRESS,
-        ):
-            for sprite, button in self.cat_sprites.items():
-                cat_id = button.return_cat_id()
-                if event.type == pygame_gui.UI_BUTTON_ON_HOVERED:
-                    if button != event.ui_element:
-                        continue
-                    self.selection_boxes[sprite].show()
-                elif (
-                    event.type == pygame_gui.UI_BUTTON_ON_UNHOVERED
-                    and cat_id not in self.selected
-                ):
-                    if button != event.ui_element:
-                        continue
-                    self.selection_boxes[sprite].hide()
-                elif event.type == pygame_gui.UI_BUTTON_START_PRESS:
-                    if button != event.ui_element:
-                        continue
-                    if cat_id in self.selected:
-                        self.selected.remove(cat_id)
-                    else:
-                        self.selected.append(cat_id)
-
-        return super().process_event(event)
-
-    def reset_selection(self):
-        for box in self.selection_boxes.values():
-            box.hide()
-        self.selected.clear()
-
-    def show(self):
-        super().show()
-
-        if self.allow_selection:
-            for sprite, button in self.cat_sprites.items():
-                cat_id = button.return_cat_id()
-                if cat_id not in self.selected:
-                    self.selection_boxes[sprite].hide()
-
-
-class UIModifiedImage(pygame_gui.elements.UIImage):
-    """
-    UIImage class modified to prevent it from blocking hover actions in other elements
-    """
-
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        image_surface: pygame.surface.Surface,
-        manager: Optional[IUIManagerInterface] = None,
-        image_is_alpha_premultiplied: bool = False,
-        container: Optional[IContainerLikeInterface] = None,
-        parent_element: Optional[UIElement] = None,
-        object_id: Optional[Union[ObjectID, str]] = None,
-        anchors: Optional[Dict[str, Union[str, UIElement]]] = None,
-        visible: int = 1,
-        *,
-        starting_height: int = 1,
-    ):
-        super().__init__(
-            relative_rect=relative_rect,
-            image_surface=image_surface,
-            manager=manager,
-            image_is_alpha_premultiplied=image_is_alpha_premultiplied,
-            container=container,
-            parent_element=parent_element,
-            object_id=object_id,
-            anchors=anchors,
-            visible=visible,
-            starting_height=starting_height,
-        )
-
-    def check_hover(self, time_delta: float, hovered_higher_element: bool) -> bool:
-        """
-        A method that helps us to determine which, if any, UI Element is currently being hovered
-        by the mouse.
-
-        :param time_delta: A float, the time in seconds between the last call to this function
-                           and now (roughly).
-        :param hovered_higher_element: A boolean, representing whether we have already hovered a
-                                       'higher' element.
-
-        :return bool: A boolean that is true if we have hovered a UI element, either just now or
-                      before this method.
-        """
-        should_block_hover = False
-        if self.alive():
-            mouse_x, mouse_y = self.ui_manager.get_mouse_position()
-            mouse_pos = pygame.math.Vector2(mouse_x, mouse_y)
-
-            if self.hover_point(mouse_x, mouse_y) and not hovered_higher_element:
-                should_block_hover = True
-
-                if self.can_hover():
-                    if not self.hovered:
-                        self.hovered = True
-                        self.on_hovered()
-
-                    self.while_hovering(time_delta, mouse_pos)
-                else:
-                    should_block_hover = False
-                    if self.hovered:
-                        self.hovered = False
-                        self.on_unhovered()
-            else:
-                if self.hovered:
-                    self.hovered = False
-                    self.on_unhovered()
-        elif self.hovered:
-            self.hovered = False
-        return should_block_hover
-
-    def can_hover(self) -> bool:
-        """
-        A stub method to override. Called to test if this method can be hovered.
-        """
-        if self.alive() and self.is_enabled:
-            return True
-        else:
-            return False
-
-
-class UIDropDown(UIDropDownContainer):
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        parent_text: str,
-        item_list: list or tuple,
-        manager: IUIManagerInterface,
-        container: UIContainer = None,
-        child_dimensions: tuple = None,
-        center_children: bool = False,
-        parent_style: ButtonStyles = ButtonStyles.DROPDOWN,
-        parent_override=None,
-        parent_reflect_selection=False,
-        child_style: ButtonStyles = ButtonStyles.DROPDOWN,
-        multiple_choice: bool = False,
-        disable_selection: bool = True,
-        starting_height: int = 1,
-        object_id: str = None,
-        visible: bool = True,
-        anchors: dict = None,
-        child_trigger_close: bool = True,
-        starting_selection: list = None,
-        open_on_hover: bool = False,
-    ):
-        """
-        Class to handle the creation and management of non-scrolling dropdowns. It's recommended to use the on_use()
-        screen func to check for changes to the selected_list attribute rather than handle_event()
-
-        :param relative_rect: The rect for the parent button, by default these dimensions are also used for the child
-        buttons. All positioning is based off this rect's position. THIS SHOULD NOT BE UI_SCALED
-        :param parent_text: The text to display on the parent button.
-        :param item_list: The list of options that will become child buttons.
-        :param child_dimensions: This overrides the relative_rect dimensions for the child buttons, allowing you to create
-        parent and child buttons with differing dimensions
-        :param center_children: Set True if child buttons should be centered beneath the parent button, rather than anchored to the parent's left side. Only useful if child dimensions are larger than the parent's. Defaults to False.
-        :param parent_style: The button style to use for the parent button, defaults to DROPDOWN
-        :param parent_override: This isn't best practice to use, but it's an exception added for the filter dropdown
-        :param parent_reflect_selection: When a selection is made, the parent text changes to reflect the selection.
-        :param child_style: The button style to use for the child buttons, defaults to DROPDOWN
-        :param multiple_choice: If the selected_list should hold multiple selections, defaults to False
-        :param disable_selection: If the clicked child_button should be disabled, defaults to True
-        :param child_trigger_close: If clicking a child_button should close the dropdown, defaults to True
-        :param starting_selection: Items from item_list that should begin selected.
-        :param open_on_hover: Dropdown will open while being hovered and close once unhovered
-        """
-        self.selected_list = (
-            [item for item in starting_selection if starting_selection]
-            if starting_selection
-            else []
-        )
-        self.multiple_choice = multiple_choice
-        self.disable_selection = disable_selection
-        self.parent_text = parent_text
-        self.parent_reflect_selection = parent_reflect_selection
-
-        super().__init__(
-            relative_rect=ui_scale(relative_rect.copy()),
-            container=container,
-            manager=manager,
-            starting_height=starting_height,
-            object_id=object_id,
-            visible=visible,
-            anchors=anchors,
-            child_trigger_close=child_trigger_close,
-            starting_selection=starting_selection,
-            open_on_hover=open_on_hover,
-        )
-
-        rect = pygame.Rect(
-            (relative_rect.x, 0), (relative_rect.width, relative_rect.height)
-        )
-
-        # create parent button
-        if not parent_override:
-            self.parent_button = UISurfaceImageButton(
-                ui_scale(rect),
-                parent_text,
-                get_button_dict(parent_style, relative_rect.size),
-                manager=manager,
-                object_id=f"@buttonstyles_{parent_style.value}",
-                container=self,
-                anchors=anchors,
-            )
-        else:
-            self.parent_button = parent_override
-            self.parent_button.set_container(self)
-
-        if center_children:
-            x_pos = -int(child_dimensions[0] / 2 - relative_rect.width / 2)
-        else:
-            x_pos = relative_rect.x
-        dropdown_rect = ((x_pos, 0), (0, 0))
-
-        self.child_button_container = UIAutoResizingContainer(
-            ui_scale(pygame.Rect(dropdown_rect)),
-            manager=manager,
-            container=self,
-            resize_left=False,
-            resize_top=False,
-            anchors=(
-                {
-                    "top_target": self.parent_button,
-                    "left_target": self.parent_button.anchors.get("left_target"),
-                }
-                if self.parent_button.anchors.get("left_target")
-                else {"top_target": self.parent_button}
-            ),
-        )
-
-        # create child buttons
-        if child_dimensions:
-            self.child_dimensions = child_dimensions
-        else:
-            self.child_dimensions = relative_rect.size
-
-        self.child_style = child_style
-
-        prev_element = None
-        self.child_button_dicts = {}
-        self.manager = manager
-
-        for child in item_list:
-            y_pos = -2 if prev_element else 0
-
-            self.child_button_dicts[child] = UISurfaceImageButton(
-                ui_scale(pygame.Rect((0, y_pos), self.child_dimensions)),
-                child,
-                get_button_dict(self.child_style, self.child_dimensions),
-                manager=manager,
-                object_id=f"@buttonstyles_{self.child_style.value}",
-                container=self.child_button_container,
-                starting_height=starting_height,
-                anchors={"top_target": prev_element} if prev_element else None,
-            )
-            prev_element = self.child_button_dicts[child]
-
-        self.child_buttons = self.child_button_dicts.values()
-        if starting_selection:
-            if disable_selection:
-                for button in starting_selection:
-                    self.child_button_dicts[button].disable()
-            if parent_reflect_selection:
-                self.parent_button.set_text(starting_selection[0])
-        self.close()
-
-    def new_item_list(self, item_list):
-        """
-        Replace the old item_list with a new one. This kills and then rebuilds the child buttons.
-        """
-        # destroy old buttons and clear selected list
-        for button in self.child_button_dicts.values():
-            button.kill()
-        self.child_button_dicts.clear()
-        self.selected_list.clear()
-
-        prev_element = None
-        for child in item_list:
-            y_pos = -2 if prev_element else 0
-
-            self.child_button_dicts[child] = UISurfaceImageButton(
-                ui_scale(pygame.Rect((0, y_pos), self.child_dimensions)),
-                child,
-                get_button_dict(self.child_style, self.child_dimensions),
-                manager=self.manager,
-                object_id=f"@buttonstyles_{self.child_style.value}",
-                container=self.child_button_container,
-                starting_height=self.starting_height,
-                anchors={"top_target": prev_element} if prev_element else None,
-            )
-            prev_element = self.child_button_dicts[child]
-
-        self.child_buttons = self.child_button_dicts.values()
-
-    def set_selected_list(self, new_list):
-        self.selected_list.clear()
-        self.selected_list = new_list
-        if self.disable_selection:
-            for item in self.selected_list:
-                self.child_button_dicts[item].disable()
-        if self.parent_reflect_selection and new_list:
-            self.parent_button.set_text(new_list[0])
-
-    def update(self, time_delta: float):
-        # updates our selection list
-        for name, button in self.child_button_dicts.items():
-            if not button.pressed:
-                continue
-
-            if self.child_trigger_close:
-                self.close()
-
-            # multiple choice
-            if self.multiple_choice:
-                if name in self.selected_list:
-                    self.selected_list.remove(name)
-                else:
-                    self.selected_list.append(name)
-
-                if self.disable_selection:
-                    button.disable()
-
-                break
-
-            # single choice
-            elif not self.multiple_choice:
-                if self.selected_list and self.selected_list[0] == name:
-                    self.selected_list.clear()
-                    if self.parent_reflect_selection:
-                        self.parent_button.set_text(self.parent_text)
-                else:
-                    self.selected_list.clear()
-                    self.selected_list.append(name)
-                    if self.parent_reflect_selection:
-                        self.parent_button.set_text(name)
-                if self.disable_selection:
-                    for other_button in self.child_buttons:
-                        other_button.enable()
-                    button.disable()
-                break
-
-        super().update(time_delta)
-
-
-class UICollapsibleContainer(
-    pygame_gui.elements.UIAutoResizingContainer, IContainerLikeInterface
-):
-    def __init__(
-        self,
-        relative_rect: RectLike,
-        title_text: str = None,
-        top_button_oriented_left: bool = True,
-        bottom_button: bool = True,
-        bottom_button_oriented_left: bool = True,
-        scrolling_container_to_reset=None,
-        min_edges_rect: pygame.Rect = None,
-        max_edges_rect: pygame.Rect = None,
-        resize_left: bool = True,
-        resize_right: bool = True,
-        resize_top: bool = True,
-        resize_bottom: bool = True,
-        manager: Optional[IUIManagerInterface] = None,
-        starting_height: int = 1,
-        container: Optional[IContainerLikeInterface] = None,
-        parent_element: Optional[UIElement] = None,
-        object_id: Optional[Union[ObjectID, str]] = None,
-        title_object_id: Optional[Union[ObjectID, str]] = None,
-        anchors: Optional[Dict[str, Union[str, UIElement]]] = None,
-        visible: int = 1,
-    ):
-        """
-        A collapsible container that can be created with a title (text visible while closed) as well as top and bottom
-        buttons on the right or left side.
-        :param title_text: Text visible while container is closed, this will align with the top button
-        :param top_button_oriented_left: The top button will appear on the far left of the container if this is True,
-        else it will appear on the right. Default is True.
-        :param bottom_button: Should this container have a bottom button. Default is True
-        :param bottom_button_oriented_left: If it has a bottom button, will it be oriented to the left side. Default is True
-        """
-        super().__init__(
-            relative_rect=relative_rect,
-            min_edges_rect=min_edges_rect,
-            max_edges_rect=max_edges_rect,
-            resize_left=resize_left,
-            resize_right=resize_right,
-            resize_top=resize_top,
-            resize_bottom=resize_bottom,
-            manager=manager,
-            starting_height=starting_height,
-            container=container,
-            parent_element=parent_element,
-            object_id=object_id,
-            anchors=anchors,
-            visible=visible,
-        )
-        self.title_text = None
-        self.top_button_oriented_left = top_button_oriented_left
-        self.bottom_button_oriented_left = bottom_button_oriented_left
-        self.scrolling_container_to_reset = scrolling_container_to_reset
-
-        rect = ui_scale(pygame.Rect((0, 0), (36, 36)))
-        if not self.top_button_oriented_left:
-            rect.topright = ((-10, 10),)
-            anchors = {"right": "right"}
-        else:
-            rect.topleft = ((10, 10),)
-            anchors = None
-
-        self.top_button = UISurfaceImageButton(
-            rect,
-            Icon.ARROW_UP,
-            get_button_dict(ButtonStyles.ICON, (36, 36)),
-            manager=manager,
-            object_id="@buttonstyles_icon",
-            starting_height=1,
-            container=self,
-            tool_tip_text="buttons.collapse_down",
-            anchors=anchors if anchors else None,
-        )
-
-        if title_text:
-            self.title_text = UITextBoxTweaked(
-                title_text,
-                ui_scale(pygame.Rect((0, 10), (-1, -1))),
-                object_id=title_object_id,
-                line_spacing=1,
-                manager=manager,
-                container=self,
-                anchors=(
-                    {"left_target": self.top_button}
-                    if self.top_button_oriented_left
-                    else None
-                ),
-            )
-
-        self.bottom_button = None
-        if bottom_button:
-            if not self.bottom_button_oriented_left:
-                rect.bottomright = ((-10, 10),)
-                anchors = {"right": "right", "bottom": "bottom"}
-            else:
-                rect.bottomleft = ((10, -10),)
-                anchors = {"bottom": "bottom"}
-
-            self.bottom_button = UISurfaceImageButton(
-                rect,
-                Icon.ARROW_UP,
-                get_button_dict(ButtonStyles.ICON, (36, 36)),
-                manager=manager,
-                object_id="@buttonstyles_icon",
-                starting_height=1,
-                container=self,
-                tool_tip_text="buttons.collapse_up",
-                anchors=anchors,
-            )
-
-        self.is_open = True
-        self.saved_scroll_position = None
-
-    def close(self):
-        """
-        Closes the container, leaving only the top button visible
-        """
-
-        for ele in self.elements:
-            if ele == self.title_text:
-                continue
-            if ele == self.top_button:
-                self.top_button.set_text(Icon.ARROW_DOWN)
-                self.top_button.set_tooltip("buttons.collapse_down")
-                continue
-            ele.hide()
-
-        self.resize_bottom = False
-        self.set_dimensions(
-            (
-                self.get_relative_rect().w,
-                self.top_button.get_relative_rect().h
-                + self.top_button.get_relative_rect().y,
-            )
-        )
-
-        # this resets the scrolling container containing this container back to its prior position (or close to it)
-        if self.scrolling_container_to_reset and self.saved_scroll_position:
-            self.scrolling_container_to_reset.scrollable_container.recalculate_abs_edges_rect()
-            self.scrolling_container_to_reset.update(1)
-
-            self.scrolling_container_to_reset.vert_scroll_bar.set_scroll_from_start_percentage(
-                self.saved_scroll_position
-            )
-            self.scrolling_container_to_reset.vert_scroll_bar.has_moved_recently = True
-            self.scrolling_container_to_reset.update(1)
-
-        self.is_open = False
-
-    def open(self):
-        """
-        Opens the container, revealing its contents
-        """
-        if self.scrolling_container_to_reset:
-            # saves the scroll positions .481 is the magic number to actually make this accurate, don't ask me why
-            self.saved_scroll_position = (
-                self.scrolling_container_to_reset.vert_scroll_bar.scroll_position
-                * 0.481
-            ) / self.scrolling_container_to_reset.vert_scroll_bar.scrollable_height
-        for ele in self.elements:
-            if ele == self.top_button:
-                self.top_button.set_text(Icon.ARROW_UP)
-                self.top_button.set_tooltip("buttons.collapse_up")
-                continue
-            ele.show()
-
-        self.resize_bottom = True
-        self.should_update_dimensions = True
-
-        self.is_open = True
-
-    def update(self, time_delta: float):
-        if self.top_button.pressed:
-            if self.is_open:
-                self.close()
-            else:
-                self.open()
-        elif self.bottom_button and self.bottom_button.pressed:
-            if self.is_open:
-                self.close()
-            else:
-                self.open()
-
-        super().update(time_delta)
