+++ scripts/screens/MedDenScreen.py
@@ -7,9 +7,7 @@ import pygame_gui
 from scripts.cat.cats import Cat
 from scripts.clan_resources.herb.herb_supply import MESSAGES
 from scripts.game_structure import game
-from scripts.game_structure.ui_elements import (
-    UIModifiedImage,
-)
+from ..ui.elements.modified_image import UIModifiedImage
 from ..ui.elements.text_box_tweaked import UITextBoxTweaked
 from ..ui.elements.sprite_button import UISpriteButton
 from ..ui.elements.image_button import UIImageButton
