+++ scripts/screens/ProfileScreen.py
@@ -14,9 +14,7 @@ from pygame_gui.core import ObjectID
 from scripts.cat.cats import Cat, BACKSTORIES
 from scripts.clan_resources.freshkill import FRESHKILL_ACTIVE
 from scripts.game_structure import image_cache, game
-from scripts.game_structure.ui_elements import (
-    UIModifiedImage,
-)
+from ..ui.elements.modified_image import UIModifiedImage
 from ..ui.elements.text_box_tweaked import UITextBoxTweaked
 from ..ui.elements.image_button import UIImageButton
 from ..ui.elements.surface_image_button import UISurfaceImageButton
