+++ scripts/ui/elements/scrolling_dropdown.py
@@ -3,9 +3,7 @@ from pygame_gui.core import UIContainer
 from pygame_gui.core.gui_type_hints import RectLike
 from pygame_gui.core.interfaces import IUIManagerInterface
 
-from scripts.game_structure.ui_elements import (
-    UIDropDownContainer,
-)
+from scripts.ui.elements.dropdown_container import UIDropDownContainer
 from scripts.ui.elements.scrolling_button_list import UIScrollingButtonList
 from scripts.ui.elements.surface_image_button import UISurfaceImageButton
 from scripts.ui.generate_button import ButtonStyles, get_button_dict
