+++ scripts/ui/windows/freshkill.py
@@ -12,10 +12,8 @@ from scripts.clan_package.settings import (
     switch_clan_setting,
 )
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import (
-    UICheckbox,
-    UICatListDisplay,
-)
+from scripts.ui.elements.cat_list_display import UICatListDisplay
+from scripts.ui.elements.checkbox import UICheckbox
 from scripts.ui.elements.modified_scrolling_container import (
     UIModifiedScrollingContainer,
 )
