+++ scripts/ui/windows/rel_change_details.py
@@ -1,18 +1,14 @@
-from msilib.schema import UIText
-
 import pygame
 import pygame_gui
 from scripts.game_structure import game
 from scripts.screens.enums import GameScreen
+from scripts.ui.elements.cat_list_display import UICatListDisplay
+from scripts.ui.elements.surface_image_button import UISurfaceImageButton
+from scripts.ui.elements.text_box_tweaked import UITextBoxTweaked
 
 from scripts.ui.windows.window_base_class import GameWindow
 from scripts.ui.scale import ui_scale, ui_scale_offset, ui_scale_value
 from scripts.game_structure.screen_settings import MANAGER
-from scripts.game_structure.ui_elements import (
-    UISurfaceImageButton,
-    UICatListDisplay,
-    UITextBoxTweaked,
-)
 from scripts.cat.cats import Cat
 from scripts.ui.generate_button import get_button_dict, ButtonStyles
 from scripts.ui.icon import Icon
